Cleaned dataset for final project
This table is based on the cleaned version of the data to ensure that there are English column names. From S24f2 only the table “Videregående uddannelser (fuldførte)” has been included, as the other tables in the spreadsheet are of no relevance for this project. This is the data on which the graphs of the project were created.